Add SignUp page error handling and change file structure
