+++ backend/App.js
@@ -1,24 +1,41 @@
-const dotenv = require("dotenv")
 const mongoose = require("mongoose")
 const express = require("express")
 const app = express()
-
-
+const dotenv = require("dotenv");
 dotenv.config({path: './config.env'})
-
-const PORT = process.env.PORT
+const cors = require("cors")
 
 //This runs the conn.js
 require('./db/conn.js')
 
+const corsOptions = {
+    origin : 'http://localhost:3000',
+    method : ['GET','POST'],
+    optionsSuccessStatus: 200
+};
+
+const port = process.env.PORT || 5000 ;
+
+// console.log("Hello");
+
 //This parses JSON automatically
 app.use(express.json());
 
+app.use(cors(corsOptions))
+
+app.use("/api",require("./routes/animeRoutes")) ;
+
+app.use("/user",require("./routes/userRoutes"))
+
+// app.get("/",(req,res)=>{ res.send("Hello World!");})
+
+
+app.listen(port,() => {
+    console.log(`server running on port ${port}`) ;
+}) ;
+
+
+
 //This include auth.js. auth.js has all the routes/posts/gets
-app.use(require('./router/auth.js'))
+// app.use(require('./router/auth.js'))
 
-
-
-app.listen(PORT, () => {
-    console.log(`server is running at port no ${PORT}`);
-})
