+++ backend/routes/userRoutes.js
@@ -6,4 +6,16 @@ const UserController = require("../controllers/userController") ;
 
 router.post("/create",UserController.createUser);
 
+
+const jwt = require('jsonwebtoken');
+// router.post('/', (req, res) => {
+//     console.log(req.body.username);
+//     res.json({ message: req.body });
+// });
+
+
+router.post('/login',UserController.loginUser)
+
+
+
 module.exports = router ;
