+++ frontend/src/Components/Login.js
@@ -22,7 +22,7 @@ export default function Login() {
     e.preventDefault()
 
 
-    const res = await fetch("/login",{
+    const res = await fetch("/user/login",{
       method: "POST",
       headers: {
         "Content-Type": "application/json"
